Move IPK installer screens out of SoftwareManager, make ipk scanning built-in
Split IPKGMenu, IPKGSource and IpkgInstaller (renamed IPKGInstaller) out
of SoftwareManager/plugin.py into their own Screens/IPKGInstaller.py, and
register the ipk file scanner as a built-in in Components/Scanner.py
instead of via a WHERE_FILESCAN plugin descriptor, decoupling ipk
handling from the SoftwareManager plugin being installed.

MediaScanner/plugin.py now uses Tools.Notifications (showInfo,
addModalNotification) instead of ChoiceBox/MessageBox directly, and
drops the InfoBar.execing guard in partitionListChanged. Since
MessageBox only returns the second element of a selected list entry
(unlike ChoiceBox, which returns the whole entry), the scan-result list
is now built as (description, (description, scanner, files, session,
popup)) pairs so execute() still receives the full tuple it expects.
