+++ lib/python/Plugins/Extensions/MediaScanner/plugin.py
@@ -3,8 +3,7 @@ from Plugins.Plugin import PluginDescriptor
 from Components.Scanner import scanDevice
 from Components.Harddisk import harddiskmanager
 from Screens.ChoiceBox import ChoiceBox
-from Screens.InfoBar import InfoBar
-from Screens.MessageBox import MessageBox
+from Tools.Notifications import showInfo, notificationCenter
 
 parentScreen = None
 global_session = None
@@ -31,17 +30,16 @@ def mountpoint_choosen(option):
 	(description, mountpoint, session, popup) = option
 	res = scanDevice(mountpoint)
 
-	list = [(r.description, r, res[r], session, popup) for r in res]
+	list = [(r.description, (r.description, r, res[r], session, popup)) for r in res]
 
 	if not list:
-		if popup:
-			if access(mountpoint, F_OK | R_OK):
-				session.open(MessageBox, _("No displayable files on this medium found!"), MessageBox.TYPE_INFO, simple=True, timeout=5)
+		if popup and access(mountpoint, F_OK | R_OK):
+			showInfo(_("No displayable files on this medium found!"))
 		if parentScreen:
 			parentScreen.close()
 		return
 
-	session.openWithCallback(execute, ChoiceBox, title=_("The following files were found..."), list=list)
+	notificationCenter.addModalNotification(_("The following files were found..."), list=list, callback=execute)
 
 
 def scan(session, parent=None):
@@ -57,10 +55,8 @@ def main(session, **kwargs):
 
 
 def partitionListChanged(action, device):
-	if InfoBar.instance:
-		if InfoBar.instance.execing:
-			if action == 'add' and device.is_hotplug:
-				mountpoint_choosen((device.description, device.mountpoint, global_session, False))
+	if action == 'add' and device.is_hotplug:
+		mountpoint_choosen((device.description, device.mountpoint, global_session, False))
 
 
 def sessionstart(reason, session):
@@ -69,12 +65,10 @@ def sessionstart(reason, session):
 
 
 def autostart(reason, **kwargs):
-	global global_session
 	if reason == 0:
 		harddiskmanager.on_partition_list_change.append(partitionListChanged)
 	elif reason == 1:
 		harddiskmanager.on_partition_list_change.remove(partitionListChanged)
-		global_session = None
 
 
 def Plugins(**kwargs):
